requirements.txt
flask
werkzeug
langchain
langchain-google-genai
python-dotenv
boto3
pypdf
tiktoken
unstructured
wikipedia
beautifulsoup4
langchain_pinecone
faiss-cpu
rank_bm25
flask_mysqldb
mysql-connector-python
-e .